rename utils into client
