+++ api/client/SearchProductApi.ts
